third_party/blink/web_tests/http/tests/dom/document-contentType-xhr-expected.txt

PASS xhr.status is 200
PASS xhr.responseXML.contentType is "text/html"
PASS xhr.status is 200
PASS xhr.responseXML.contentType is "text/xml"
PASS xhr.status is 200
PASS xhr.responseXML.contentType is "application/xml"
PASS xhr.status is 200
PASS xhr.responseXML.contentType is "text/html"
PASS xhr.status is 200
PASS xhr.responseXML.contentType is "text/xml"
PASS xhr.status is 200
PASS xhr.responseXML.contentType is "application/xml"
PASS successfullyParsed is true

TEST COMPLETE
